Also search for a share_dir in a venv-specific path
When running in a venv, sys.prefix is modified to reflect the venv's
location whereas sysconfig paths are not.  So we need to use sys.prefix
as part of one of the search paths to find the share dir when running in
a venv.

We still check the sysconfig datarootdir path because this can be set by
the user when Python is built.  So outside of a venv it might be
something user-defined instead of sys.prefix/share/.