Rework share logic a bit to prevent TypeError
Change:
Previously when constructing the tuple of possible paths to search, if
any of the `sysconfig.get_config_var(...)` returned None (which it can),
we would bomb out instead of just skipping that path.

Instead encapsulate all the logic of finding the share directory into a
function and use that. We handle the previous possible TypeError by
ensuring that each use of get_config_var() does not return None.

This also sates mypy and makes types happy.

Test Plan:
- CI (tox tests+mypy, functional smoke tests)

Signed-off-by: Rick Elrod <rick@elrod.me>