[yaml] remove explicit sort_keys (#542)
[yaml] remove explicit sort_keys

Change:

Remove explicit sort_keys which got passed to yaml.dump(). Old PyYAML did
not allow this kwarg. Old and new PyYAML both default to sorting keys, and
we never override that, so just remove the line and let PyYAML do its thing.
If we need to override it in the future, we can add a version check (though
it gets annoying to handle version comparisons manually).

Test Plan:

Tested locally on PyYAML 3 and PyYAML 5.4.

Tickets:

Fixes #517
Fixes #522
Fixes #525

Signed-off-by: Rick Elrod rick@elrod.me

Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: None <None>