Replace use of yum w/ rpm (#614)
Replace use of yum w/ rpm

Fixes #613
EEs (downstream) no longer have yum installed, so switch to RPM for package information.
The shared parser could not easily be adapted to the way RPM formats it's output, so parse the RPM output in a dedicated parser
as always, the real changes are in the image introspection script, the majority of the fixture updates are a back log from before the default EE was changed.

Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: None <None>