Move the flake8 config into a standalone location (#683)
Move the flake8 config into a standalone location

The motivation here is to improve the maintainability and follow the wider community practices. It also allows to find out more easily where the common flake8 configuration exists rather than trying to check all of the config files where it may be present.
As a bonus, this will allow comparing/syncing flake8 configurations across projects that also follow this practice.

Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: None <None>