Pin `astroid` due to false `no-name-in-module` and `no-member` errors (#773)
Pin `astroid` due to false `no-name-in-module` and `no-member` errors

astroid >= 2.9.1, =< 2.9.3 (a pylint dependency) has a bug causing the reports of false-positive no-name-in-module and no-member rule violations.
This patch pins the respective transitive dependency to the last known good version of 2.9.0 as a temporary workaround.
It should be reverted as soon as the underlying upstream bug is fixed.
Ref: PyCQA/pylint#5131
Related: #772

Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: None <None>