Rename `Entry` in the `configuration_subsystem` to `SettingsEntry` (#774)
Rename `Entry` in the `configuration_subsystem` to `SettingsEntry`

Also EntryValue to SettingsEntryValue

update code as needed
fix for new mypy error resulting from type in post processor

This should eliminate a subset of the docs errors, as Entry was defined in the tokenization subsystem as well.
The use of settings vs configuration was done as over time the use of configuration within navigator should slowly be replace with settings to avoid confusion with ansible.

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: None <None>