Add trailing commas (#788)
Changes to make the use of trailing commas consistent throughout the project

This modifies the changed files such that the use of trailing commas is consistent through out the project.
Which in turn allows all C812 exceptions to be removed from the flake8 configuration file which are a result of a linter being vetted called flake8-commas which is part of the wemake style guide.
Although this does not enforce the use of trailing commas, it will allow users of tox -e lint-vetting to use it to both add trailing commas where they are missing and see errors that indicate a missing trailing comma which would be inconsistent with the repository.
The changes were made by the precommit formatter called add-trailing-comma followed by black.  Only one file needed to be manually editted to break a long string.
tests/integration/_tmux_session.py:277:101: E501 line too long (101 > 100 characters)
                    f"******** ERROR: TMUX '{err_message}' TIMEOUT @ {elapsed}s @ {tstamp} ********",

(the string was simply broken in half)

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: None <None>