change (#791)
Small formatting change for comment in setup.py

Make a small formatting change to the comment in setup.py.
This was necessary to eliminate an error thrown (E800) by flake8-eradicate which is one linter being vetted.
By reducing the errors thrown by tox -e lint-vetting it should allow future contributors to use it with needing to look through long lists of errors unrelated to their code, and see feedback related only to the changes being made in a PR.

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: None <None>