double quotes (#787)
Changes making the use of double quotes consistent throughout

Changes in 2 files to switch to double quotes from single

which in turn allows for:
This allows the flake8-quotes inline-quote style to be a double quote, eliminating the individual Q000 entries throughout the flake8 config file.
flake8-quotes is a linter pulled in by the we-make-styleguide which is a linter being vetted.
Although this change does not enforce the use of double quotes, it will allow users of the tox -e lint-vetting to see warning messages that double quotes are use consistently throughout the repo and a single quote is being introduced.
Some good conversation here about this: https://issuehunt.io/r/wemake-services/wemake-python-styleguide/issues/681
Closes #757

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: None <None>