Sort the entries in the flake8 `per_file_ignore` list (#794)
Sort the entries in the flake8 `per_file_ignore` list

Sort the entries in the flake8 per_file_ingore list and add a note asking they are sorted in the future.
This should make it easier to find and add entries in the future in a predictable spot.
No changes were made to any line, just re-odered.

Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: None <None>