enable_isort (#795)
Enable automatic import statement sorting

This enables a pre-commit hook to sort all import statements across the project so future contributors do not need to order them in a manner that is consistent with the repository. This should save all of us time and brain cycles during submissions and review.
Because pre-commit and isort will run automatically for every pull-request, it doesn't even need to be run locally as part of a developer workflow. (but can be with tox -e lint, so the changes can be seen locally)

Remove the stages: ['manual'] from the precommit entry, which will allow it to run when using tox -e lint
Add flake8-isort as a required check after isort runs.
Update contributors guide to reflect the use of isort

(this is a trust and verify approach, trust isort will do it's job, verify it did with the flake8 extention)
This PR is not intended to revisit the current configuration, please log and submit issues and PRs in the future for this.

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: Rick Elrod <rick@elrod.me>
Reviewed-by: Alison Hart <contact@alisonlhart.com>
Reviewed-by: Jake Jackson  <jljacks93@gmail.com>
Reviewed-by: Priyam Sahoo <None>
Reviewed-by: Ganesh Nalawade <None>
Reviewed-by: None <None>