Add MVP GHA workflows (#650)
Add MVP GHA workflows

This change aims to add a MVP implementation of GHA that replicates current zuul config.
Deliberately it does not include features that are not already covered by zuul jobs, and it also does not include a release pipeline.
The look_nots are being removed because /home/runner was used as a check in those tests to confirm that the test was not running in an EE.   Because GHA uses runner as the default user in it's images, we cannot use this test in the future.

Reviewed-by: Sviatoslav Sydorenko <webknjaz+github/profile@redhat.com>
Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: Rick Elrod <rick@elrod.me>
Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: None <None>