Form button validation, pass a `List` not `Generator` (#790)
Form button validation, pass a `List` not `Generator`

Form buttons should really only have one of 2 validators, either all_true in the case of a Submit button, or none in the case of a Cancel button.
This changes results in passing a List to the button's conditional validator rather than a Generator.
Several typing updates to correct what is being passed through.
This also eliminates a  trailing-comma error reported when using tox -e lint-vetting

Reviewed-by: Sorin Sbarnea <sorin.sbarnea@gmail.com>
Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: None <None>