Activate `pylint-pytest` plugin (#782)
Activate `pylint-pytest` plugin

This patch is necessary because pylint does not autoactivate third-
party plugins just because they are installed in the same virtualenv.
It seems like this bit of the original configuration got lost over
time: the Python distribution package is present in the dependencies
and gets installed but pylint does not use it. With this change, it
becomes used again.

Reviewed-by: Bradley A. Thornton <bthornto@redhat.com>
Reviewed-by: None <None>