Add more vars to tox passenv
This adds a list of variables that we want to pass to tox
environments, some of them being critical for functioning of podman
and docker. Fixes the problem that prevented use of remote container
engines with tox.

The list was based on what we use on other similar projects, like
molecule-podman and molecule-docker.