Indroduce an intermediate lint-candidates target (#825)
This change exposes an additional vetting-like linting environment
that is limited to rules that are actively being addressed by the
contributors. This rules sub-list is to be kept small so that it
could be promoted put to voting and made enforced easily. Once the
rules are in a good shape, a new focused list of related rules can
take their place and become linting candidates.

Co-authored-by: Bradley A. Thornton <bthornto@redhat.com>