Exterminate black poison from the checks
This is necessary because it was never accepted by all DevTools
contributors and imposing it on unsuspecting victims is not fair.
The contribution guide says that it's not enforced but it's not
entirely correct -- it is not possible to contribute without having
to reproduce the same result as black emits.

This is undesirable because it puts extra burden on contributors who
want to see their code structured in a meaningful way. While black
helps with the style consistency, it's not the only way of achieving
it. And in its current state, it sometimes makes formatting changes
that don't make sense and work against readability. Unfortunately,
because of it's repressive nature, it's not possible to fix it while
having it enabled.

Black also hides a number of useful linting violations that, when
actually shown, can help the coders identify structural/architectural
problems.

As for alternatives, it's possible to maintain consistent code style
in a friendlier way by combining several other smaller formatters
(like autopep8, some configurations of yapf and pre-commit.com hooks)
that are as acceptable PEP 8-wise but give a bit more flexibility in
preserving structural readability of contributions. This is not to
say that people are now allowed to nitpick style in reviews but it
should be noted that sometimes important structural improvement
requests get mistaken for style change. When getting reviews, the
motivation of improving structure should be assumed, if the
suggested code snippets contains style updates, it can be disgregarded
for as long as the underlying idea is well understood and considered
by the PR submitter.