Fix warning message issue for run action in stdout mode
Fixes https://github.com/ansible/ansible-navigator/issues/849

In case of multiline warnings messages the ansible-runner emits multiple
`verbose` event dicitonary with the same `id`. The runner event are added
in the `queue` by navigator and since the `id` of event is same it picks
up the last event and duplicates the event causing duplicate warnings.